+++ e2e/e2etests/test_bitcoin_withdraw_multiple.go
@@ -1,10 +1,6 @@
 package e2etests
 
 import (
-	"math/big"
-	"strconv"
-
-	"github.com/btcsuite/btcutil"
 	"github.com/stretchr/testify/require"
 
 	"github.com/zeta-chain/zetacore/e2e/runner"
@@ -33,20 +29,4 @@ func WithdrawBitcoinMultipleTimes(r *runner.E2ERunner, args []string) {
 	for i := 0; i < times; i++ {
 		withdrawBTCZRC20(r, receiver, amount)
 	}
-}
-
-func parseInt(t require.TestingT, s string) int {
-	v, err := strconv.Atoi(s)
-	require.NoError(t, err, "unable to parse int from %q", s)
-
-	return v
-}
-
-// bigIntFromFloat64 takes float64 (e.g. 0.001) that represents btc amount
-// and converts it to big.Int for downstream usage.
-func btcAmountFromFloat64(t require.TestingT, amount float64) *big.Int {
-	satoshi, err := btcutil.NewAmount(amount)
-	require.NoError(t, err)
-
-	return big.NewInt(int64(satoshi))
 }
