+++ e2e/scripts/debug.sh
